+++ scandeval/benchmarks/nordial.py
@@ -1,4 +1,4 @@
-'''Nynorsk classification of a language model on the NorDial dataset'''
+'''Dialect classification of a language model on the NorDial dataset'''
 
 from datasets import Dataset
 from typing import Tuple
@@ -44,11 +44,12 @@ class NorDialBenchmark(TextClassificationBenchmark):
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
                  verbose: bool = False):
-        id2label = ['bokmål', 'nynorsk']
+        id2label = ['bokmål', 'nynorsk', 'dialectal', 'mixed']
         label_synonyms = [
             ['LABEL_0', id2label[0]],
             ['LABEL_1', id2label[1]],
-            ['LABEL_2', id2label[2]]
+            ['LABEL_2', id2label[2]],
+            ['LABEL_3', id2label[3]]
         ]
         super().__init__(name='nordial',
                          id2label=id2label,